Prevent assessor database to be be faked in production

+++ lib/tasks/assessor.rake
@@ -1,6 +1,10 @@
 desc 'Import some random assessors data'
 
 task :generate_assessor do
+  if ENV['STAGE'] == 'production'
+    exit
+  end
+
   ActiveRecord::Base.connection.execute('TRUNCATE TABLE assessors RESTART IDENTITY')
 
   result = ActiveRecord::Base.connection.execute('SELECT * FROM postcode_geolocation ORDER BY random() LIMIT 1000')
